+++ features/public-site/components/index.ts
@@ -1,0 +1,3 @@
+export { PrivacyPolicyPage } from "./privacy-policy-page";
+export { PublicHomePage } from "./public-home-page";
+export { TermsOfServicePage } from "./terms-of-service-page";